Issue #257 fix tests
