+++ service/src/main/java/com/myhome/repositories/AmenityBookingItemRepository.java
@@ -7,8 +7,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.util.HashSet;
 import java.util.Optional;
 
 public interface AmenityBookingItemRepository extends JpaRepository<AmenityBookingItem, String> {
     Optional<AmenityBookingItem> findByAmenityBookingItemId(String amenityBookingItemId);
+
+    HashSet<AmenityBookingItem> findAllByAmenity_AmenityId(String amenityId);
 }
