+++ service/src/main/java/com/myhome/services/AmenityService.java
@@ -30,9 +30,6 @@ public interface AmenityService {
 
   boolean deleteAmenity(String amenityId);
 
-  Set<Amenity> listAllAmenities(String communityId);
-
   boolean updateAmenity(AmenityDto updatedAmenityDto);
 
-  boolean deleteBooking(String bookingId);
 }
